upd: rearranged CLI commands for better maintenance
