+++ malevich/_utility/package.py
@@ -49,7 +49,7 @@ class PackageManager(metaclass=SingletonMeta):
             raise Exception(f"Package {package_name} does not exist")
         return package_path
 
-    def get_all_packages(self) -> None:
+    def get_all_packages(self) -> list[str]:
         return [
             package
             for package in os.listdir(self.__malevich_path)
